BFD-4705: Change metrics namespace per env and combine request logs
