+++ apps/bfd-server-ng/src/main/java/gov/cms/bfd/server/ng/interceptor/RequestMetricsInterceptor.java
@@ -5,10 +5,14 @@ import static gov.cms.bfd.server.ng.util.LoggerConstants.*;
 import ca.uhn.fhir.interceptor.api.Hook;
 import ca.uhn.fhir.interceptor.api.Interceptor;
 import ca.uhn.fhir.interceptor.api.Pointcut;
+import ca.uhn.fhir.rest.api.server.RequestDetails;
+import ca.uhn.fhir.rest.server.exceptions.BaseServerResponseException;
 import gov.cms.bfd.server.ng.log.RequestTelemetryLogger;
 import lombok.RequiredArgsConstructor;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.Bundle;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
 
 /** HAPI FHIR interceptor that records request lifecycle timing metrics. */
@@ -16,6 +20,8 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class RequestMetricsInterceptor {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(RequestMetricsInterceptor.class);
 
   private final RequestTelemetryLogger requestTelemetryLogger;
 
@@ -31,10 +37,15 @@ public class RequestMetricsInterceptor {
     requestTelemetryLogger.recordTimestamp(HAPI_INCOMING_POST_PROCESS);
   }
 
-  /** Pointcut to log timestamp in milliseconds when a request is pre-handled. */
+  /**
+   * Pointcut to log timestamp in milliseconds when a request is pre-handled.
+   *
+   * @param requestDetails the request details
+   */
   @Hook(Pointcut.SERVER_INCOMING_REQUEST_PRE_HANDLED)
-  public void requestPreHandled() {
+  public void requestPreHandled(RequestDetails requestDetails) {
     requestTelemetryLogger.recordTimestamp(HAPI_INCOMING_PRE_HANDLE);
+    requestTelemetryLogger.recordRequestDetails(requestDetails);
   }
 
   /**
@@ -61,5 +72,15 @@ public class RequestMetricsInterceptor {
   @Hook(Pointcut.SERVER_PROCESSING_COMPLETED)
   public void serverProcessCompleted() {
     requestTelemetryLogger.recordTimestamp(HAPI_PROCESS_COMPLETED);
+  }
+
+  /**
+   * Pointcut to log exceptions.
+   *
+   * @param exception exception
+   */
+  @Hook(Pointcut.SERVER_HANDLE_EXCEPTION)
+  public void handleException(BaseServerResponseException exception) {
+    requestTelemetryLogger.logRequestException(exception);
   }
 }
